document-processing-agent/requirements.txt
khaos-agent>=1.0.0
autogen-agentchat>=0.7.0
anthropic>=0.30.0
